Move proprietary EAP proxy into prebuilt lib
The eap_proxy_qmi.c is the only file with dependency on proprietary
headers. Move it into a separate shared library, so that the rest of
wpa_supplicant can be open source and built from sources.

eap_proxy_qmi.c and other parts of wpa_supplicant have circular
depenencies. To resolve this, introduce a wpa_supplicant_base library
with code that eap_proxy_qmi.c depends on, but that does itself not call
into the proxy eap_proxy code.

Issue: FP3-A11#226
Change-Id: I713bdb9ef123858f7fee59e49ae10603a50b365a